Move test to /unit
